🔧 BD322N code: rename pr meta type model
