+++ packages/agentplane/src/commands/shared/pr-meta/model.ts
